Create reserve (#7)
* createReserveScreen showSearchPassenger animate_do

* Provider w/ formz - Create/Update Reserve

A provider was created to control the information on the frontend
to send it to the database. In turn, this same provider, with some
modifications, will work to update the data. It now remains to
design the fields to complete before sending the information to
the API.

* search passenger delegate

* serviceType, startdate, startTime, tripType

* continued create Form

* call to API

The API call is made and the new reservation is saved. It remains to solve the enterprise_id (bring it from the database to send it)

* form finished

* kotlin version, refresh data, filter and fix bug

Auth0 requires kotlin 1.9.0, added RefreshIndicator to refresh data
on providers and screend, added filter on reserve screen and
fix a bug on the stream of all searchDelegates

* reserve detail fix, search reserve, reserve form

reserve detail fix on interpriseId null, search deletage reserve
hint item w/ custom_slide, reserve form dropdownmenu according to
figma, modularization of custom slide